refactor: clean and refactor frontend code.
