+++ frontend/src/app/(user)/(main)/layout.tsx
@@ -1,7 +1,9 @@
 import { Inter } from "next/font/google";
-import Header from "@/components/Header";
 import { AuthProvider } from "@/hooks/auth";
 import "@/styles/globals.css";
+import Header from "@/components/Header";
+import { userNavItems } from "@/constants";
+import { UserRound } from "lucide-react";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -13,7 +15,11 @@ export default function UserLayout({
     return (
         <html lang="en" className="h-full">
             <body className={`${inter.className} flex flex-col h-full`}>
-                <Header />
+                <Header
+                    headerContent="User"
+                    navItems={userNavItems}
+                    icon={<UserRound size={32} />}
+                />
                 <AuthProvider>
                     <div className="flex-grow">{children}</div>
                 </AuthProvider>
