+++ frontend/src/app/(user)/dashboard/layout.tsx
@@ -1,7 +1,9 @@
 import { Inter } from "next/font/google";
-import Header from "@/components/dashboard/Header";
 import "@/styles/globals.css";
 import { AuthProvider } from "@/hooks/auth";
+import { UserRound } from "lucide-react";
+import { dashboardNavItems } from "@/constants";
+import Header from "@/components/Header";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -13,7 +15,11 @@ export default function UserLayout({
     return (
         <html lang="en" className="h-full">
             <body className={`${inter.className} flex flex-col h-full`}>
-                <Header />
+                <Header
+                    headerContent="User"
+                    navItems={dashboardNavItems}
+                    icon={<UserRound size={32} />}
+                />
                 <AuthProvider>
                     <div className="flex-grow">{children}</div>
                 </AuthProvider>
