+++ frontend/src/components/Header.tsx
@@ -4,20 +4,23 @@ import React from "react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { setCookie } from "cookies-next";
-import { message, Button } from "antd";
-import { Settings } from "lucide-react";
+import { message } from "antd";
+import { Button } from "antd";
 import { useRouter } from "next/navigation";
+import { INavItem } from "@/types";
 
-const Header = () => {
+const Header = ({
+    icon,
+    headerContent,
+    navItems,
+}: {
+    icon: React.ReactNode;
+    headerContent: string;
+    navItems: INavItem[];
+}) => {
     const [messageApi, contextHolder] = message.useMessage();
     const router = useRouter();
     const pathname = usePathname();
-
-    const navItems = [
-        { name: "Knowledge Base", path: "/knowledge" },
-        { name: "Chat", path: "/chat" },
-        { name: "Dashboard", path: "/dashboard" },
-    ];
 
     const successMessage = ({
         content,
@@ -49,7 +52,7 @@ const Header = () => {
 
     const isActive = (path: string) => {
         if (pathname === "/") return false;
-        return pathname.startsWith(path);
+        return pathname == path;
     };
 
     const handleLogout = async () => {
@@ -71,18 +74,20 @@ const Header = () => {
             <header className="bg-white shadow-sm h-16">
                 <div className="container mx-auto px-4 h-full flex items-center justify-between">
                     <div className="flex items-center space-x-4">
-                        <Settings size={32} />
-                        <span className="font-bold text-2xl">User</span>
+                        {icon}
+                        <span className="font-bold text-xl">
+                            {headerContent}
+                        </span>
                     </div>
                     <nav className="flex space-x-4">
                         {navItems.map((item) => (
                             <Link
                                 key={item.path}
                                 href={item.path}
-                                className={`px-3 py-2 relative ${
+                                className={`px-3 py-2 relative font-semibold text-sm md:text-lg ${
                                     isActive(item.path)
                                         ? "text-blue-500 font-bold after:content-[''] after:absolute after:bottom-0 after:left-0 after:w-full after:h-[2px] after:bg-blue-500"
-                                        : "text-gray-600 hover:bg-blue-500 hover:text-white hover:after:content-[''] hover:after:absolute hover:after:bottom-0 hover:after:left-0 hover:after:w-full hover:after:h-[2px] hover:after:bg-white hover:rounded-lg"
+                                        : "text-gray-600 hover:bg-blue-500 hover:text-white hover:after:content-[''] hover:after:absolute hover:after:bottom-0 hover:after:left-0 hover:after:w-full hover:after:h-[2px] hover:after:bg-white hover:rounded-lg transition-all duration-300"
                                 }`}
                             >
                                 {item.name}
@@ -91,11 +96,10 @@ const Header = () => {
                     </nav>
                     <div className="flex items-center space-x-10">
                         <Button
-                            type="primary"
-                            className="bg-red-500 ml-auto"
+                            className="ml-auto border-3 text-red-500 shadow-md border-red-300"
                             onClick={handleLogout}
                         >
-                            Logout
+                            Đăng xuất
                         </Button>
                     </div>
                 </div>
