+++ frontend/src/components/chat/SideBar.tsx
@@ -3,9 +3,8 @@ import { message } from "antd";
 import { Plus, HelpCircle } from "lucide-react";
 import React, { useCallback, useRef, useState, useEffect } from "react";
 
-import { IAssistant } from "@/app/(user)/(main)/chat/page";
-import { IConversation } from "@/app/(user)/(main)/chat/page";
 import ConversationCard from "./ConversationCard";
+import { IAssistant, IConversation } from "@/types";
 
 const Sidebar = ({
     isVisible,
