--- frontend/src/types/auth.ts
@@ -1,8 +1,0 @@
-export interface ILoginFormValues {
-    username: string;
-    password: string;
-}
-
-export interface ILoginResponse {
-    access_token: string;
-}
